requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pillow==10.4.0
transformers==4.44.2
safetensors==0.4.4
